[Refactor] add product seeder
change type data price in table product

+++ database/migrations/2025_01_07_033904_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->float('price');
+            $table->decimal('price', 15, 2);
             $table->foreignId('category_id')->constrained('product_categories');
             $table->foreignId('sub_category_id')->constrained('product_sub_categories');
             $table->timestamps();
